Add new features and fix linting issues
