+++ src/index.ts
@@ -1,6 +1,7 @@
 #!/usr/bin/env node
 
 import type { Provider } from '@nimiq/utils/fiat-api'
+import type { CliConfig } from './utils.js'
 import process from 'node:process'
 import { Server } from '@modelcontextprotocol/sdk/server/index.js'
 import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
@@ -17,6 +18,7 @@ import { calculateStakingRewards } from '@nimiq/utils/rewards-calculator'
 import { posSupplyAt } from '@nimiq/utils/supply-calculator'
 import MiniSearch from 'minisearch'
 import { initRpcClient } from 'nimiq-rpc-client-ts/client'
+
 import {
   getAccountByAddress,
   getActiveValidators,
@@ -32,161 +34,10 @@ import {
   getValidators,
   isConsensusEstablished,
 } from 'nimiq-rpc-client-ts/http'
-import * as v from 'valibot'
-
-interface CliConfig {
-  rpcUrl: string
-  rpcUsername?: string
-  rpcPassword?: string
-}
-
-// Valibot schemas for input validation
-const SupplyAtSchema = v.object({
-  timestampMs: v.pipe(v.number(), v.description('The timestamp in milliseconds at which to calculate the PoS supply')),
-  network: v.optional(v.pipe(v.picklist(['main-albatross', 'test-albatross']), v.description('The network name')), 'main-albatross'),
-})
-
-const StakingRewardsSchema = v.object({
-  stakedSupplyRatio: v.optional(v.pipe(v.number(), v.description('The ratio of the total staked cryptocurrency to the total supply. If not provided, the current staked ratio will be used instead.'))),
-  amount: v.optional(v.pipe(v.number(), v.description('The initial amount of cryptocurrency staked, in NIM')), 1),
-  days: v.optional(v.pipe(v.number(), v.description('The number of days the cryptocurrency is staked')), 365),
-  autoRestake: v.optional(v.pipe(v.boolean(), v.description('Indicates whether the staking rewards are restaked')), true),
-  network: v.optional(v.pipe(v.picklist(['main-albatross', 'test-albatross']), v.description('The network name')), 'main-albatross'),
-  fee: v.optional(v.pipe(v.number(), v.description('The fee percentage that the pool charges for staking')), 0),
-})
-
-const PriceSchema = v.object({
-  currencies: v.pipe(v.array(v.string()), v.description('An array of currency tickers to get the price against (e.g., ["USD", "EUR", "BTC"])')),
-  provider: v.optional(v.pipe(v.picklist(['CryptoCompare', 'CoinGecko']), v.description('The provider to use for fetching prices')), 'CryptoCompare'),
-})
-
-const BlockByNumberSchema = v.object({
-  blockNumber: v.pipe(v.number(), v.description('The block number to retrieve')),
-  includeBody: v.optional(v.pipe(v.boolean(), v.description('Whether to include the block body with transactions')), false),
-})
-
-const BlockByHashSchema = v.object({
-  hash: v.pipe(v.string(), v.description('The block hash to retrieve')),
-  includeBody: v.optional(v.pipe(v.boolean(), v.description('Whether to include the block body with transactions')), false),
-})
-
-const AccountSchema = v.object({
-  address: v.pipe(v.string(), v.description('The Nimiq address to get account information for')),
-  withMetadata: v.optional(v.pipe(v.boolean(), v.description('Whether to include additional metadata')), false),
-})
-
-const TransactionSchema = v.object({
-  hash: v.pipe(v.string(), v.description('The transaction hash to retrieve')),
-})
-
-const TransactionsByAddressSchema = v.object({
-  address: v.pipe(v.string(), v.description('The Nimiq address to get transactions for')),
-  max: v.optional(v.pipe(v.number(), v.description('Maximum number of transactions to return')), 100),
-  startAt: v.optional(v.pipe(v.string(), v.description('Transaction hash to start at, used for paging'))),
-  onlyConfirmed: v.optional(v.pipe(v.boolean(), v.description('Whether to only return confirmed transactions')), true),
-})
-
-const ValidatorsSchema = v.object({
-  includeStakers: v.optional(v.pipe(v.boolean(), v.description('Whether to include staker information for each validator')), false),
-  onlyActive: v.optional(v.pipe(v.boolean(), v.description('If true, returns only active validators. If false, returns all validators.')), true),
-})
-
-const ValidatorSchema = v.object({
-  address: v.pipe(v.string(), v.description('The validator address to get information for')),
-})
-
-const SlotsSchema = v.object({
-  blockNumber: v.optional(v.pipe(v.number(), v.description('Block number to get slots for (optional, defaults to current)'))),
-})
-
-const HeadSchema = v.object({
-  includeBody: v.optional(v.pipe(v.boolean(), v.description('Whether to include the block body with transactions')), false),
-})
-
-const RpcMethodsSchema = v.object({
-  includeSchemas: v.optional(v.pipe(v.boolean(), v.description('Whether to include parameter and result schemas for each method')), false),
-})
-
-const SearchDocsSchema = v.object({
-  query: v.pipe(v.string(), v.description('The search query to find relevant documentation sections. Use 2-4 specific keywords for best results (e.g., "transaction fee", "validator rewards", "web client setup")')),
-  limit: v.optional(v.pipe(v.number(), v.description('Maximum number of search results to return (default: 10)')), 10),
-})
-
-// Utility function for validating inputs with Valibot
-function validateInput<T>(schema: v.GenericSchema<T>, input: unknown): T {
-  try {
-    const result = v.parse(schema, input)
-    return result
-  }
-  catch (error) {
-    if (v.isValiError(error)) {
-      const issues = v.flatten(error.issues)
-      const errorMessages = Object.entries(issues.nested || {}).map(([key, value]) =>
-        `${key}: ${Array.isArray(value) ? value.join(', ') : value}`,
-      )
-      throw new McpError(
-        ErrorCode.InvalidParams,
-        `Invalid input parameters: ${errorMessages.join('; ')}`,
-      )
-    }
-    throw new McpError(ErrorCode.InvalidParams, 'Invalid input parameters')
-  }
-}
-
-// Parse CLI arguments
-function parseArgs(): CliConfig {
-  // In Cloudflare Workers, process.argv is undefined, so return default config
-  if (typeof process === 'undefined' || !process.argv) {
-    return {
-      rpcUrl: 'https://rpc.nimiqwatch.com',
-      rpcUsername: undefined,
-      rpcPassword: undefined,
-    }
-  }
-
-  const args = process.argv.slice(2)
-  const config: CliConfig = {
-    rpcUrl: 'https://rpc.nimiqwatch.com',
-    rpcUsername: undefined,
-    rpcPassword: undefined,
-  }
-
-  for (let i = 0; i < args.length; i++) {
-    switch (args[i]) {
-      case '--rpc-url':
-        config.rpcUrl = args[++i]
-        break
-      case '--rpc-username':
-        config.rpcUsername = args[++i]
-        break
-      case '--rpc-password':
-        config.rpcPassword = args[++i]
-        break
-      case '--help':
-      case '-h':
-        console.error(`
-Nimiq MCP Server
-
-Usage: nimiq-mcp [options]
-
-Options:
-  --rpc-url <url>         Nimiq RPC endpoint URL (default: https://rpc.nimiqwatch.com)
-  --rpc-username <user>   RPC username for authentication (optional)
-  --rpc-password <pass>   RPC password for authentication (optional)
-  --help, -h              Show this help message
-
-Examples:
-  nimiq-mcp
-  nimiq-mcp --rpc-url https://rpc.nimiqwatch.com
-  nimiq-mcp --rpc-url https://localhost:8648 --rpc-username user --rpc-password pass
-`)
-        process.exit(0)
-        break
-    }
-  }
-
-  return config
-}
+import * as schemas from './schemas.js'
+import { TOOL_DEFINITIONS } from './tool-definitions.js'
+// Import shared modules
+import { buildElicitationPrompt, createSnippet, parseArgs, validateInput, VERSION } from './utils.js'
 
 export class NimiqMcpServer {
   public server: Server
@@ -196,12 +47,12 @@ export class NimiqMcpServer {
   private cachedDocs: string | null = null
 
   constructor(config?: CliConfig) {
-    this.config = config || this.parseArgs()
+    this.config = config || parseArgs()
 
     this.server = new Server(
       {
         name: 'nimiq-mcp',
-        version: '1.0.0',
+        version: VERSION,
       },
       {
         capabilities: {
@@ -214,11 +65,6 @@ export class NimiqMcpServer {
     this.setupToolHandlers()
     this.setupResourceHandlers()
     this.setupErrorHandling()
-  }
-
-  // Make parseArgs protected so it can be overridden
-  protected parseArgs(): CliConfig {
-    return parseArgs()
   }
 
   // Make this method public so it can be called from Worker
@@ -242,344 +88,7 @@ export class NimiqMcpServer {
   public setupToolHandlers(): void {
     this.server.setRequestHandler(ListToolsRequestSchema, async () => {
       return {
-        tools: [
-          {
-            name: 'get_nimiq_supply',
-            description: 'Get the current supply of NIM. Returns an object with the following supply fields: total, vested, burned, max, initial, staking, minted, circulating, and mined. All values are in Luna (1 NIM = 100,000 Luna).',
-            inputSchema: {
-              type: 'object',
-              properties: {},
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'calculate_nimiq_supply_at',
-            description: 'Calculate the Nimiq PoS supply at a given time',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                timestampMs: {
-                  type: 'number',
-                  description: 'The timestamp in milliseconds at which to calculate the PoS supply',
-                },
-                network: {
-                  type: 'string',
-                  enum: ['main-albatross', 'test-albatross'],
-                  description: 'The network name',
-                  default: 'main-albatross',
-                },
-              },
-              required: ['timestampMs'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'calculate_nimiq_staking_rewards',
-            description: 'Calculates the potential wealth accumulation based on staking. If `stakedSupplyRatio` is not provided, the current staked ratio will be used instead.',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                stakedSupplyRatio: {
-                  type: 'number',
-                  description: 'The ratio of the total staked cryptocurrency to the total supply. If not provided, the current staked ratio will be used instead.',
-                },
-                amount: {
-                  type: 'number',
-                  description: 'The initial amount of cryptocurrency staked, in NIM',
-                  default: 1,
-                },
-                days: {
-                  type: 'number',
-                  description: 'The number of days the cryptocurrency is staked',
-                  default: 365,
-                },
-                autoRestake: {
-                  type: 'boolean',
-                  description: 'Indicates whether the staking rewards are restaked',
-                  default: true,
-                },
-                network: {
-                  type: 'string',
-                  enum: ['main-albatross', 'test-albatross'],
-                  description: 'The network name',
-                  default: 'main-albatross',
-                },
-                fee: {
-                  type: 'number',
-                  description: 'The fee percentage that the pool charges for staking',
-                  default: 0,
-                },
-              },
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_price',
-            description: 'Get the price of NIM against other currencies',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                currencies: {
-                  type: 'array',
-                  items: {
-                    type: 'string',
-                  },
-                  description: 'An array of currency tickers to get the price against (e.g., ["USD", "EUR", "BTC"])',
-                },
-                provider: {
-                  type: 'string',
-                  enum: ['CryptoCompare', 'CoinGecko'],
-                  description: 'The provider to use for fetching prices',
-                  default: 'CryptoCompare',
-                },
-              },
-              required: ['currencies'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_head',
-            description: 'Get the current head block of the Nimiq blockchain',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                includeBody: {
-                  type: 'boolean',
-                  description: 'Whether to include the block body with transactions',
-                  default: false,
-                },
-              },
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_block_by_number',
-            description: 'Get a Nimiq block by its number',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                blockNumber: {
-                  type: 'number',
-                  description: 'The block number to retrieve',
-                },
-                includeBody: {
-                  type: 'boolean',
-                  description: 'Whether to include the block body with transactions',
-                  default: false,
-                },
-              },
-              required: ['blockNumber'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_block_by_hash',
-            description: 'Get a Nimiq block by its hash',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                hash: {
-                  type: 'string',
-                  description: 'The block hash to retrieve',
-                },
-                includeBody: {
-                  type: 'boolean',
-                  description: 'Whether to include the block body with transactions',
-                  default: false,
-                },
-              },
-              required: ['hash'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_account',
-            description: 'Get Nimiq account information by address',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                address: {
-                  type: 'string',
-                  description: 'The Nimiq address to get account information for',
-                },
-                withMetadata: {
-                  type: 'boolean',
-                  description: 'Whether to include additional metadata',
-                  default: false,
-                },
-              },
-              required: ['address'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_balance',
-            description: 'Get the balance of a Nimiq account',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                address: {
-                  type: 'string',
-                  description: 'The Nimiq address to get balance for',
-                },
-                withMetadata: {
-                  type: 'boolean',
-                  description: 'Whether to include additional metadata',
-                  default: false,
-                },
-              },
-              required: ['address'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_transaction',
-            description: 'Get Nimiq transaction details by hash',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                hash: {
-                  type: 'string',
-                  description: 'The transaction hash to retrieve',
-                },
-              },
-              required: ['hash'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_transactions_by_address',
-            description: 'Get Nimiq transactions for a specific address',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                address: {
-                  type: 'string',
-                  description: 'The Nimiq address to get transactions for',
-                },
-                max: {
-                  type: 'number',
-                  description: 'Maximum number of transactions to return',
-                  default: 100,
-                },
-                startAt: {
-                  type: 'string',
-                  description: 'Transaction hash to start at, used for paging',
-                },
-                onlyConfirmed: {
-                  type: 'boolean',
-                  description: 'Whether to only return confirmed transactions',
-                  default: true,
-                },
-              },
-              required: ['address'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_validators',
-            description: 'Get information about Nimiq validators. By default, it returns only active validators.',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                includeStakers: {
-                  type: 'boolean',
-                  description: 'Whether to include staker information for each validator',
-                  default: false,
-                },
-                onlyActive: {
-                  type: 'boolean',
-                  description: 'If true, returns only active validators. If false, returns all validators.',
-                  default: true,
-                },
-              },
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_validator',
-            description: 'Get Nimiq validator information by address',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                address: {
-                  type: 'string',
-                  description: 'The validator address to get information for',
-                },
-              },
-              required: ['address'],
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_slots',
-            description: 'Get Nimiq validator slots information',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                blockNumber: {
-                  type: 'number',
-                  description: 'Block number to get slots for (optional, defaults to current)',
-                },
-              },
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_epoch_number',
-            description: 'Get the current Nimiq epoch number',
-            inputSchema: {
-              type: 'object',
-              properties: {},
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_network_info',
-            description: 'Get Nimiq network information including peer count and sync status',
-            inputSchema: {
-              type: 'object',
-              properties: {},
-              additionalProperties: false,
-            },
-          },
-          {
-            name: 'get_nimiq_rpc_methods',
-            description: 'Get all available Nimiq RPC methods from the latest OpenRPC document',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                includeSchemas: {
-                  type: 'boolean',
-                  description: 'Whether to include parameter and result schemas for each method',
-                  default: false,
-                },
-              },
-              additionalProperties: false,
-            },
-          },
-
-          {
-            name: 'search_nimiq_docs',
-            description: 'Search through the Nimiq documentation using full-text search. For best results, use specific keywords rather than full sentences (e.g., "validator staking" instead of "how do I stake with a validator")',
-            inputSchema: {
-              type: 'object',
-              properties: {
-                query: {
-                  type: 'string',
-                  description: 'The search query to find relevant documentation sections. Use 2-4 specific keywords for best results (e.g., "transaction fee", "validator rewards", "web client setup")',
-                },
-                limit: {
-                  type: 'number',
-                  description: 'Maximum number of search results to return (default: 10)',
-                  default: 10,
-                },
-              },
-              required: ['query'],
-              additionalProperties: false,
-            },
-          },
-        ],
+        tools: TOOL_DEFINITIONS,
       }
     })
 
@@ -629,6 +138,8 @@ export class NimiqMcpServer {
 
           case 'search_nimiq_docs':
             return await this.handleSearchDocs(args)
+          case 'interactive_staking_calculator':
+            return await this.handleInteractiveStakingCalculator(args)
 
           default:
             throw new McpError(
@@ -675,25 +186,42 @@ export class NimiqMcpServer {
   }
 
   private handleCalculateSupplyAt(args: any): any {
-    const validatedInput = validateInput(SupplyAtSchema, args)
+    const validatedInput = validateInput(schemas.SupplyAtSchema, args)
     const { timestampMs, network } = validatedInput
-    const supply = posSupplyAt(timestampMs, { network })
-    return {
-      content: [
-        {
-          type: 'text',
-          text: JSON.stringify({
-            timestampMs,
-            network,
-            supply,
-          }, null, 2),
-        },
-      ],
+
+    try {
+      const supply = posSupplyAt(timestampMs, { network })
+      const timestamp = new Date(timestampMs).toISOString()
+
+      return {
+        content: [
+          {
+            type: 'text',
+            text: JSON.stringify({
+              timestampMs,
+              timestamp,
+              network,
+              supply: {
+                value: supply,
+                unit: 'Luna',
+                nim: supply / 100000, // Convert to NIM for readability
+              },
+              calculatedAt: new Date().toISOString(),
+            }, null, 2),
+          },
+        ],
+      }
+    }
+    catch (error: any) {
+      throw new McpError(
+        ErrorCode.InternalError,
+        `Failed to calculate supply at timestamp ${timestampMs}: ${error.message}`,
+      )
     }
   }
 
   private async handleCalculateStakingRewards(args: any): Promise<any> {
-    const validatedInput = validateInput(StakingRewardsSchema, args)
+    const validatedInput = validateInput(schemas.StakingRewardsSchema, args)
     const newArgs = { ...validatedInput }
 
     if (newArgs.stakedSupplyRatio === undefined || newArgs.stakedSupplyRatio === null) {
@@ -727,7 +255,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetNimPrice(args: any): Promise<any> {
-    const validatedInput = validateInput(PriceSchema, args)
+    const validatedInput = validateInput(schemas.PriceSchema, args)
     const { currencies, provider } = validatedInput
     const typedProvider = provider as Provider
     const vsCurrencies = currencies.map((c: string) => c.toUpperCase())
@@ -770,7 +298,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetHead(args: any): Promise<any> {
-    const validatedInput = validateInput(HeadSchema, args)
+    const validatedInput = validateInput(schemas.HeadSchema, args)
     const { includeBody } = validatedInput
 
     // Get the current block number (head)
@@ -815,7 +343,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetBlockByNumber(args: any): Promise<any> {
-    const validatedInput = validateInput(BlockByNumberSchema, args)
+    const validatedInput = validateInput(schemas.BlockByNumberSchema, args)
     const { blockNumber, includeBody } = validatedInput
 
     const [success, error, block] = await getBlockByNumber(
@@ -845,7 +373,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetBlockByHash(args: any): Promise<any> {
-    const validatedInput = validateInput(BlockByHashSchema, args)
+    const validatedInput = validateInput(schemas.BlockByHashSchema, args)
     const { hash, includeBody } = validatedInput
 
     const [success, error, block] = await getBlockByHash(
@@ -875,7 +403,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetAccount(args: any): Promise<any> {
-    const validatedInput = validateInput(AccountSchema, args)
+    const validatedInput = validateInput(schemas.AccountSchema, args)
     const { address, withMetadata } = validatedInput
 
     const params: any = {}
@@ -906,7 +434,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetBalance(args: any): Promise<any> {
-    const validatedInput = validateInput(AccountSchema, args)
+    const validatedInput = validateInput(schemas.AccountSchema, args)
     const { address, withMetadata } = validatedInput
 
     const params: any = {}
@@ -940,7 +468,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetTransaction(args: any): Promise<any> {
-    const validatedInput = validateInput(TransactionSchema, args)
+    const validatedInput = validateInput(schemas.TransactionSchema, args)
     const { hash } = validatedInput
 
     const [success, error, transaction] = await getTransactionByHash({ hash }, {})
@@ -967,7 +495,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetTransactionsByAddress(args: any): Promise<any> {
-    const validatedInput = validateInput(TransactionsByAddressSchema, args)
+    const validatedInput = validateInput(schemas.TransactionsByAddressSchema, args)
     const { address, max, startAt, onlyConfirmed } = validatedInput
 
     const params: any = { address, max }
@@ -1004,7 +532,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetValidators(args: any): Promise<any> {
-    const validatedInput = validateInput(ValidatorsSchema, args)
+    const validatedInput = validateInput(schemas.ValidatorsSchema, args)
     const { includeStakers, onlyActive } = validatedInput
 
     const params: any = {}
@@ -1067,7 +595,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetValidator(args: any): Promise<any> {
-    const validatedInput = validateInput(ValidatorSchema, args)
+    const validatedInput = validateInput(schemas.ValidatorSchema, args)
     const { address } = validatedInput
 
     const [success, error, validator] = await getValidatorByAddress({ address }, {})
@@ -1094,7 +622,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetSlots(args: any): Promise<any> {
-    const validatedInput = validateInput(SlotsSchema, args)
+    const validatedInput = validateInput(schemas.SlotsSchema, args)
     const { blockNumber } = validatedInput
 
     if (!blockNumber) {
@@ -1215,7 +743,7 @@ export class NimiqMcpServer {
   }
 
   private async handleGetRpcMethods(args: any): Promise<any> {
-    const validatedInput = validateInput(RpcMethodsSchema, args)
+    const validatedInput = validateInput(schemas.RpcMethodsSchema, args)
     const { includeSchemas } = validatedInput
 
     try {
@@ -1416,7 +944,7 @@ export class NimiqMcpServer {
 
   private async handleSearchDocs(args: any): Promise<any> {
     try {
-      const validatedInput = validateInput(SearchDocsSchema, args)
+      const validatedInput = validateInput(schemas.SearchDocsSchema, args)
       const { query, limit } = validatedInput
 
       // Initialize search index if not already done
@@ -1436,7 +964,7 @@ export class NimiqMcpServer {
         section: result.section,
         score: result.score,
         // Include a snippet of the content around matches
-        snippet: this.createSnippet(result.content, query),
+        snippet: createSnippet(result.content, query),
       }))
 
       return {
@@ -1461,34 +989,89 @@ export class NimiqMcpServer {
     }
   }
 
-  private createSnippet(content: string, query: string, maxLength: number = 200): string {
-    const words = content.toLowerCase().split(/\s+/)
-    const queryWords = query.toLowerCase().split(/\s+/)
+  private async handleInteractiveStakingCalculator(args: any): Promise<any> {
+    try {
+      const { amount, days, autoRestake, network = 'main-albatross' } = args
 
-    // Find the first occurrence of any query word
-    let startIndex = -1
-    for (let i = 0; i < words.length; i++) {
-      if (queryWords.some(qWord => words[i].includes(qWord))) {
-        startIndex = Math.max(0, i - 10) // Start 10 words before the match
-        break
+      // Check which parameters are missing and need elicitation
+      const missingParams: string[] = []
+      if (amount === undefined || amount === null)
+        missingParams.push('amount')
+      if (days === undefined || days === null)
+        missingParams.push('days')
+      if (autoRestake === undefined || autoRestake === null)
+        missingParams.push('autoRestake')
+
+      // If we have missing parameters, use elicitation to request them
+      if (missingParams.length > 0) {
+        const elicitationPrompt = buildElicitationPrompt(missingParams)
+
+        return {
+          content: [
+            {
+              type: 'text',
+              text: JSON.stringify({
+                status: 'elicitation_required',
+                message: 'Some parameters are missing for the staking calculation. Please provide the following information:',
+                missingParameters: missingParams,
+                prompt: elicitationPrompt,
+                currentParameters: { amount, days, autoRestake, network },
+                example: 'You can call this tool again with: interactive_staking_calculator {"amount": 1000, "days": 365, "autoRestake": true}',
+              }, null, 2),
+            },
+          ],
+        }
+      }
+
+      // All parameters provided, calculate staking rewards
+      const stakingArgs: any = { amount, days, autoRestake, network }
+
+      // Get current staked supply ratio
+      try {
+        const supplyData = await this.getSupplyData()
+        if (supplyData.circulating > 0) {
+          stakingArgs.stakedSupplyRatio = supplyData.staking / supplyData.circulating
+        }
+      }
+      catch {
+        // Use default ratio if we can't fetch supply data
+        stakingArgs.stakedSupplyRatio = 0.3 // Default 30% staked ratio
+      }
+
+      const rewards = calculateStakingRewards(stakingArgs)
+
+      return {
+        content: [
+          {
+            type: 'text',
+            text: JSON.stringify({
+              status: 'calculation_complete',
+              parameters: {
+                amount: `${amount} NIM`,
+                stakingPeriod: `${days} days`,
+                autoRestake: autoRestake ? 'Yes' : 'No',
+                network,
+                stakedSupplyRatio: `${(stakingArgs.stakedSupplyRatio * 100).toFixed(2)}%`,
+              },
+              results: {
+                initialAmount: `${(rewards as any).amount || amount} NIM`,
+                finalAmount: `${(rewards as any).finalAmount || 'N/A'} NIM`,
+                totalRewards: `${(rewards as any).totalRewards || 'N/A'} NIM`,
+                apr: `${((rewards as any).apr * 100 || 0).toFixed(2)}%`,
+                apy: `${((rewards as any).apy * 100 || 0).toFixed(2)}%`,
+              },
+              calculatedAt: new Date().toISOString(),
+            }, null, 2),
+          },
+        ],
       }
     }
-
-    if (startIndex === -1) {
-      // No direct match found, return beginning of content
-      startIndex = 0
+    catch (error) {
+      throw new McpError(
+        ErrorCode.InternalError,
+        `Failed to calculate staking rewards: ${error instanceof Error ? error.message : String(error)}`,
+      )
     }
-
-    // Get a snippet around the match
-    const snippetWords = words.slice(startIndex, startIndex + 40) // About 40 words
-    let snippet = snippetWords.join(' ')
-
-    // Truncate if too long
-    if (snippet.length > maxLength) {
-      snippet = `${snippet.substring(0, maxLength)}...`
-    }
-
-    return snippet
   }
 
   // Make this method public so it can be called from Worker
@@ -1645,8 +1228,8 @@ export class NimiqMcpServer {
   }
 }
 
-// Start the server only if not in Cloudflare Workers environment
-if (typeof process !== 'undefined' && process.argv && require.main === module) {
+// Start the server only if not in Cloudflare Workers environment and this is the main module
+if (typeof process !== 'undefined' && process.argv && import.meta.url === `file://${process.argv[1]}`) {
   const server = new NimiqMcpServer()
   server.run().catch((error) => {
     console.error('Failed to start server:', error)
